more moving around of things
